% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{asType,Array,character-method}
\alias{asType,Array,character-method}
\title{asType}
\usage{
\S4method{asType}{Array,character}(a, type)
}
\arguments{
\item{a}{Khiva array.}

\item{type}{Desired type.}
}
\description{
Returns a copy of the KHIVA array with the specified base type.
}
