% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{asType}
\alias{asType}
\title{asType}
\usage{
asType(a, type)
}
\arguments{
\item{a}{Khiva array.}

\item{type}{Desired type.}
}
\description{
Returns a copy of the KHIVA array with the specified base type.
}
