% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{AutoCorrelation}
\alias{AutoCorrelation}
\title{AutoCorrelation}
\usage{
AutoCorrelation(arr, max.lag, unbiased)
}
\arguments{
\item{arr}{KHIVA array with the time series.}

\item{max.lag}{The maximum lag to compute.}

\item{unbiased}{Determines whether it divides by n - lag (if true) or n (if false).}
}
\value{
The autocorrelation value for the given time series.
}
\description{
Calculates the autocorrelation of the specified lag for the given time series.
}
