% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{BinnedEntropy}
\alias{BinnedEntropy}
\title{BinnedEntropy}
\usage{
BinnedEntropy(arr, max.bins)
}
\arguments{
\item{arr}{KHIVA array with the time series.}

\item{max.bins}{The number of bins.}
}
\value{
The binned entropy value for the given time series.
}
\description{
Calculates the binned entropy for the given time series and number of bins.
}
