% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{bitShiftL}
\alias{bitShiftL}
\title{bitShiftL}
\usage{
bitShiftL(e1, n)
}
\arguments{
\item{e1}{Khiva array.}

\item{n}{Bits to shift.}
}
\description{
Shifts each element of the input KHIVA array n positions to left.
}
