% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{bitShiftR}
\alias{bitShiftR}
\title{bitShifR}
\usage{
bitShiftR(e1, n)
}
\arguments{
\item{e1}{Khiva array.}

\item{n}{bits to shift.}
}
\description{
Shifts each element of the input KHIVA array n positions to right.
}
