% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{CountBelowMean}
\alias{CountBelowMean}
\title{CountBelowMean}
\usage{
CountBelowMean(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
The number of values in the time series that are lower than the mean.
}
\description{
Calculates the number of values in the time series that are lower than
the mean.
}
