% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{CovarianceStatistics}
\alias{CovarianceStatistics}
\title{CovarianceStatistics}
\usage{
CovarianceStatistics(tss, unbiased)
}
\arguments{
\item{tss}{Expects an input array whose dimension zero is the length of the time series (all the same) and
dimension one indicates the number of time series.}

\item{unbiased}{Determines whether it divides by n -1 (if false) or n (if true).}
}
\value{
The covariance matrix of the time series.
}
\description{
Returns the covariance matrix of the time series contained in tss.
}
