% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{CrossCorrelation}
\alias{CrossCorrelation}
\title{CrossCorrelation}
\usage{
CrossCorrelation(arr.xss, arr.yss, unbiased)
}
\arguments{
\item{arr.xss}{KHIVA array with the time series.}

\item{arr.yss}{KHIVA array with the time series.}

\item{unbiased}{Determines whether it divides by n - lag (if true) or n (if false).}
}
\value{
The cross-correlation value for the given time series.
}
\description{
Calculates the cross-correlation of the given time series.
}
