% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{FftCoefficient}
\alias{FftCoefficient}
\title{FftCoefficient}
\usage{
FftCoefficient(arr, coefficient)
}
\arguments{
\item{arr}{KHIVA array with the time series.}

\item{coefficient}{The coefficient to extract from the FFT.}
}
\value{
KHIVA Array with:
real: The real part of the coefficient.
imag: The imaginary part of the coefficient.
abs: The absolute value of the coefficient.
angle: The angle of the coefficient.
}
\description{
Calculates the fourier coefficients of the one-dimensional discrete
Fourier Transform for real input by fast fourier transformation algorithm.
}
