% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{FindBestNDiscords}
\alias{FindBestNDiscords}
\title{FindBestNDiscords}
\usage{
FindBestNDiscords(profile, index, m, n, selfJoin)
}
\arguments{
\item{profile}{KHIVA Array with the matrix profile containing the minimum distance of each
subsequence.}

\item{index}{KHIVA Array with the matrix profile index containing where each minimum occurs.}

\item{m}{Subsequence length value used to calculate the input matrix profile.}

\item{n}{Number of motifs to extract.}

\item{selfJoin}{Indicates whether the input profile comes from a self join operation or not. It determines 
whether the mirror similar region is included in the output or not.}
}
\value{
A list of KHIVA Arrays with the discord distances, the discord indices and the subsequences indices.
}
\description{
This function extracts the best N discords from a previously calculated matrix profile.
}
