% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{GetBackendInfo}
\alias{GetBackendInfo}
\title{GetBackendInfo}
\usage{
GetBackendInfo()
}
\value{
Information of the active backend.
}
\description{
Returns a string with information from the active backend.
}
