% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{getBoolean}
\alias{getBoolean}
\title{getBoolean}
\usage{
getBoolean(a)
}
\arguments{
\item{a}{Array of bits.}
}
\value{
Logical R array with the R logical equivalent to a.
}
\description{
Gets the logical R value from C++ Boolean data type.
}
