% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{getCols,Array,numeric,numeric-method}
\alias{getCols,Array,numeric,numeric-method}
\title{getCols}
\usage{
\S4method{getCols}{Array,numeric,numeric}(a, first, last)
}
\description{
Returns a sequence of columns limited by first and last, both included.
}
