% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{getCols}
\alias{getCols}
\title{getCols}
\usage{
getCols(a, first, last)
}
\arguments{
\item{a}{Khiva Array.}

\item{first}{Index of the first column.}

\item{last}{Index of the second column.}
}
\description{
Returns a sequence of columns limited by first and last, both included.
}
