% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{getRow,Array,numeric-method}
\alias{getRow,Array,numeric-method}
\title{getRow}
\usage{
\S4method{getRow}{Array,numeric}(a, index)
}
\arguments{
\item{a}{Khiva array.}

\item{index}{Row index.}
}
\description{
Returns the row specified by index.
}
