% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{getRows,Array,numeric,numeric-method}
\alias{getRows,Array,numeric,numeric-method}
\title{getRows}
\usage{
\S4method{getRows}{Array,numeric,numeric}(a, first, last)
}
\arguments{
\item{a}{Khiva Array.}

\item{first}{Index of the first row.}

\item{last}{Index of the last row.}
}
\description{
Returns a sequence of rows limited by first and last, both included.
}
