% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{getRows}
\alias{getRows}
\title{getRows}
\usage{
getRows(a, first, last)
}
\arguments{
\item{a}{Khiva Array.}

\item{first}{Index of the first row.}

\item{last}{Index of the last row.}
}
\description{
Returns a sequence of rows limited by first and last, both included.
}
