% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{>,Array,Array-method}
\alias{>,Array,Array-method}
\title{Compares (element-wise) if an Array is greater than other.}
\usage{
\S4method{>}{Array,Array}(e1, e2)
}
\description{
Compares (element-wise) if an Array is greater than other.
}
