% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{HasDuplicates}
\alias{HasDuplicates}
\title{HasDuplicates}
\usage{
HasDuplicates(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
Array containing True if the time series contains duplicated elements
and false otherwise.
}
\description{
Calculates if the input time series contain duplicated elements.
}
