% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{IndexMassQuantile}
\alias{IndexMassQuantile}
\title{IndexMassQuantile}
\usage{
IndexMassQuantile(arr, q)
}
\arguments{
\item{arr}{KHIVA array with the time series.}

\item{q}{The quantile.}
}
\value{
The index of the mass quantile q.
}
\description{
Calculates the index of the mass quantile.
}
