% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{KHIVABackendFromOrdinal}
\alias{KHIVABackendFromOrdinal}
\title{Gets the Backend from an ordinal.}
\usage{
KHIVABackendFromOrdinal(number)
}
\arguments{
\item{number}{The backend ordinal.}
}
\description{
Gets the Backend from an ordinal.
}
