% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{KMeans}
\alias{KMeans}
\title{KMeans}
\usage{
KMeans(tss, k, tolerance = 1e-10, maxIterations = 100)
}
\arguments{
\item{tss}{Expects an input array whose dimension zero is the length of the time series (all the same) and
dimension one indicates the number of time series.}

\item{k}{The number of means to be computed.}

\item{tolerance}{The error tolerance to stop the computation of the centroids.}

\item{max_iterations}{The maximum number of iterations allowed.}
}
\value{
List with the kMeans centroids as first element and kMeans labels as second element.
}
\description{
Calculates the K-Means algorithm.
}
\details{
[1] S. Lloyd. 1982. Least squares quantization in PCM. IEEE Transactions on Information Theory, 28, 2,
Pages 129-137.
}
