% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{LargeStandardDeviation}
\alias{LargeStandardDeviation}
\title{LargeStandardDeviation}
\usage{
LargeStandardDeviation(arr, r)
}
\arguments{
\item{arr}{KHIVA array with the time series.}

\item{r}{The threshold.}
}
\value{
Array containing True for those time series in arr that have a large standard deviation.
}
\description{
Checks if the time series within arr have a large standard deviation.
}
