% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{LastLocationOfMaximum}
\alias{LastLocationOfMaximum}
\title{LastLocationOfMaximum}
\usage{
LastLocationOfMaximum(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
The last relative location of the maximum value of each series.
}
\description{
Calculates the last location of the maximum value of each time series. The position
is calculated relatively to the length of the series.
}
