% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{Mass}
\alias{Mass}
\title{Mueen's Algorithm for Similarity Search.}
\usage{
Mass(query, tss)
}
\arguments{
\item{query}{KHIVA Array whose first dimension is the length of the query time series and the second dimension is the
number of queries.}

\item{tss}{KHIVA Array whose first dimension is the length of the time series and the second dimension is the number of
time series.}
}
\value{
KHIVA Array with the distances.
}
\description{
The result has the following structure:
- 1st dimension corresponds to the index of the subsequence in the time series.
- 2nd dimension corresponds to the number of queries.
- 3rd dimension corresponds to the number of time series.
}
\details{
For example, the distance in the position (1, 2, 3) correspond to the distance of the third query to the fourth time
series for the second subsequence in the time series.

[1] Chin-Chia Michael Yeh, Yan Zhu, Liudmila Ulanova, Nurjahan Begum, Yifei Ding, Hoang Anh Dau, Diego Furtado Silva,
Abdullah Mueen, Eamonn Keogh (2016). Matrix Profile I: All Pairs Similarity Joins for Time Series: A Unifying View
that Includes Motifs, Discords and Shapelets. IEEE ICDM 2016.
}
