% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{matMult,Array,Array-method}
\alias{matMult,Array,Array-method}
\title{matMult}
\usage{
\S4method{matMult}{Array,Array}(e1, e2)
}
\arguments{
\item{e1}{Khiva array.}

\item{e2}{Khiva array.}
}
\description{
Performs a matrix multiplication.
}
