% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{MaxMinNorm}
\alias{MaxMinNorm}
\title{MaxMinNorm}
\usage{
MaxMinNorm(tss, high = 1, low = 1, epsilon = 1e-08)
}
\arguments{
\item{tss}{KHIVA array with the time series.}

\item{high}{Maximum final value (Defaults to 1.0).}

\item{low}{Minimum final value (Defaults to 0.0).}

\item{epsilon}{Safeguard for constant (or near constant) time series as the operation implies a unit scale
operation between min and max values in the tss.}
}
\value{
: KHIVA array with the same dimensions as tss where the time series have been adjusted for zero mean and
one as standard deviation.
}
\description{
ormalizes the given time series according to its minimum and maximum value and adjusts each value within the
range [low, high].
}
