% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Mean}
\alias{Mean}
\title{Mean}
\usage{
Mean(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
The mean value of each time series within arr.
}
\description{
Calculates the mean value for each time series within arr.
}
