% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{MeanNorm}
\alias{MeanNorm}
\title{MeanNorm}
\usage{
MeanNorm(tss)
}
\arguments{
\item{tss}{KHIVA array with the time series.}
}
\value{
An array with the same dimensions as tss, whose values (time series in dimension 0) have been
normalized by substracting the mean from each number and dividing each number by \eqn{max(x) - min(x)}, in the
time series.
}
\description{
Normalizes the given time series according to its maximum-minimum value and its mean. It follows the following
formulae:
\deqn{ \acute{x} = \frac{x - mean(x)}{max(x) - min(x)}.}
}
