% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Median}
\alias{Median}
\title{Median}
\usage{
Median(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
The median value of each time series within arr.
}
\description{
Calculates the median value for each time series within arr.
}
