% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Minimum}
\alias{Minimum}
\title{Minimum}
\usage{
Minimum(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
The minimum value of each time series within arr.
}
\description{
Calculates the minimum value for each time series within arr.
}
