% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{not.khiva,Array-method}
\alias{not.khiva,Array-method}
\title{not.khiva}
\usage{
\S4method{not.khiva}{Array}(e1)
}
\arguments{
\item{e1}{Khiva array.}
}
\description{
Returns the complement of the input boolean KHIVA array.
}
