% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{SampleEntropy}
\alias{SampleEntropy}
\title{SampleEntropy}
\usage{
SampleEntropy(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
An array with the same dimensions as arr, whose values (time series in dimension 0)
contains the vectorized sample entropy for all the input time series in arr.
}
\description{
Calculates a vectorized sample entropy algorithm.
https://en.wikipedia.org/wiki/Sample_entropy
https://www.ncbi.nlm.nih.gov/pubmed/10843903?dopt=Abstract
For short time-series this method is highly dependent on the parameters, but should be stable for N > 2000,
see: Yentes et al. (2012) - The Appropriate Use of Approximate Entropy and Sample Entropy with Short Data Sets
Other shortcomings and alternatives discussed in:
Richman & Moorman (2000) - Physiological time-series analysis using approximate entropy and sample entropy.
}
