% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{SampleStdevStatistics}
\alias{SampleStdevStatistics}
\title{SampleStdevStatistics}
\usage{
SampleStdevStatistics(tss)
}
\arguments{
\item{tss}{Expects an input array whose dimension zero is the length of the time series (all the same) and
dimension one indicates the number of time series.}
}
\value{
The sample standard deviation.
}
\description{
Estimates standard deviation based on a sample. The standard deviation is calculated using the "n-1" method.
}
