% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{Sbd}
\alias{Sbd}
\title{SBD}
\usage{
Sbd(tss)
}
\arguments{
\item{tss}{Expects an input array whose dimension zero is the length of the time series (all the same) and
dimension one indicates the number of time series.}
}
\value{
KHIVA array with an upper triangular matrix where each position corresponds to the distance between two
time series. Diagonal elements will be zero. For example: Position row 0 column 1 records the distance
between time series 0 and time series 1.
}
\description{
Calculates the Shape-Based distance (SBD). It computes the normalized cross-correlation and it returns 1.0 
minus the value that maximizes the correlation value between each pair of time series.
}
