% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Skewness}
\alias{Skewness}
\title{Skewness}
\usage{
Skewness(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
Array containing the skewness of each time series in arr.
}
\description{
Calculates the sample skewness of arr (calculated with the adjusted Fisher-Pearson standardized
moment coefficient G1).
}
