% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{/,Array,Array-method}
\alias{/,Array,Array-method}
\title{Divides two arrays.}
\usage{
\S4method{/}{Array,Array}(e1, e2)
}
\arguments{
\item{e1}{Khiva array.}

\item{e2}{Khiva array.}
}
\description{
Divides two arrays.
}
