% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{StompSelfJoin}
\alias{StompSelfJoin}
\title{StompSelfJoin}
\usage{
StompSelfJoin(t, m)
}
\arguments{
\item{t}{KHIVA Array the time series.}

\item{m}{KHIVA Array the time series.}
}
\value{
List of KHIVA Arrays with the matrix profile and the index profile
}
\description{
STOMP algorithm to calculate the matrix profile between 't' and itself using a subsequence length
of 'm'. This method filters the trivial matches.
}
\details{
[1] Yan Zhu, Zachary Zimmerman, Nader Shakibay Senobari, Chin-Chia Michael Yeh, Gareth Funning, Abdullah Mueen, Philip Brisk and 
Eamonn Keogh (2016). Matrix Profile II: Exploiting a Novel Algorithm and GPUs to break the one Hundred Million Barrier for 
Time Series Motifs and Joins. IEEE ICDM 2016.
}
