% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{SumOfReoccurringValues}
\alias{SumOfReoccurringValues}
\title{SumOfReoccurringValues}
\usage{
SumOfReoccurringValues(arr, is.sorted = FALSE)
}
\arguments{
\item{arr}{KHIVA array with the time series.}

\item{is.sorted}{Indicates if the input time series is sorted or not. Defaults to false.}
}
\value{
: KHIVA array with the sum of all values, that are present in the time series more than once.
}
\description{
Calculates the sum of all values, that are present in the time series more than once.
}
