% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{transpose,Array,logical-method}
\alias{transpose,Array,logical-method}
\title{transpose}
\usage{
\S4method{transpose}{Array,logical}(a, conjugate)
}
\arguments{
\item{a}{Khiva Array.}

\item{conjugate}{Indicates if the tranpose is conjugated or not.}
}
\description{
Transpose the KHIVA Array.
}
