% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{transpose}
\alias{transpose}
\title{transpose}
\usage{
transpose(a, conjugate)
}
\arguments{
\item{a}{Khiva Array.}

\item{conjugate}{Indicates if the tranpose is conjugated or not.}
}
\description{
Transpose the KHIVA Array.
}
