% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{ValueCount}
\alias{ValueCount}
\title{ValueCount}
\usage{
ValueCount(arr, v)
}
\arguments{
\item{arr}{KHIVA array with the time series.}

\item{v}{The value to be counted.}
}
\value{
An array containing the count of the given value in each time series.
}
\description{
Counts occurrences of value in the time series arr.
}
