% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Variance}
\alias{Variance}
\title{Variance}
\usage{
Variance(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
KHIVA array containing the variance in each time series.
}
\description{
Computes the variance for the time series array.
}
