% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{xor.khiva}
\alias{xor.khiva}
\title{xor.khiva}
\usage{
xor.khiva(e1, e2)
}
\arguments{
\item{e1}{Khiva array.}

\item{e2}{Khiva array.}
}
\description{
Returns a exclusive-or between each pair of elements from two KHIVA arrays (element-wise).
}
