% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{xor.khiva,Array,Array-method}
\alias{xor.khiva,Array,Array-method}
\title{xor.khiva}
\usage{
\S4method{xor.khiva}{Array,Array}(e1, e2)
}
\description{
Returns a exclusive-or between each pair of elements from two KHIVA arrays (element-wise).
}
