% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{FindBestNDiscords}
\alias{FindBestNDiscords}
\title{FindBestNDiscords}
\usage{
FindBestNDiscords(profile, index, n)
}
\arguments{
\item{profile}{KHIVA Array with the matrix profile containing the minimum distance of each
subsequence.}

\item{index}{KHIVA Array with the matrix profile index containing where each minimum occurs.}

\item{n}{Number of motifs to extract.}
}
\value{
A list of KHIVA Arrays with the discord distances, the discord indices and the subsequences indices.
}
\description{
This function extracts the best N discords from a previously calculated matrix profile.
}
